--- tests/integration/test_wazuh_db/README.md
@@ -1,2 +1,0 @@
-# Wazuh DB
-This tests are aimed to test all the functionality of WazuhDB Daemon.
